requirements.txt
Pillow>=10.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
